+++ mithril-aggregator/src/tools/mocks.rs
@@ -2,7 +2,7 @@ use async_trait::async_trait;
 
 use mithril_cardano_node_chain::chain_observer::{ChainObserver, ChainObserverError};
 use mithril_cardano_node_chain::entities::{ChainAddress, TxDatum};
-use mithril_common::crypto_helper::{KESPeriod, OpCert};
+use mithril_common::crypto_helper::{KesPeriod, OpCert};
 use mithril_common::entities::{ChainPoint, Epoch, StakeDistribution};
 use mithril_persistence::store::StakeStorer;
 
@@ -32,7 +32,7 @@ mock! {
         async fn get_current_kes_period(
             &self,
             opcert: &OpCert,
-        ) -> Result<Option<KESPeriod>, ChainObserverError>;
+        ) -> Result<Option<KesPeriod>, ChainObserverError>;
     }
 }