+++ mithril-common/src/protocol/signer_builder.rs
@@ -1,12 +1,13 @@
+use std::sync::Arc;
+
 use anyhow::Context;
 use rand_chacha::ChaCha20Rng;
 use rand_core::{CryptoRng, RngCore, SeedableRng};
-use std::path::Path;
 use thiserror::Error;
 
 use crate::{
     crypto_helper::{
-        ProtocolAggregateVerificationKey, ProtocolClerk, ProtocolClosedKeyRegistration,
+        KesSigner, ProtocolAggregateVerificationKey, ProtocolClerk, ProtocolClosedKeyRegistration,
         ProtocolInitializer, ProtocolKeyRegistration, ProtocolStakeDistribution,
     },
     entities::{PartyId, ProtocolParameters, SignerWithStake},
@@ -90,12 +91,12 @@ impl SignerBuilder {
     fn build_single_signer_with_rng<R: RngCore + CryptoRng>(
         &self,
         signer_with_stake: SignerWithStake,
-        kes_secret_key_path: Option<&Path>,
+        kes_signer: Option<Arc<dyn KesSigner>>,
         rng: &mut R,
     ) -> StdResult<(SingleSigner, ProtocolInitializer)> {
         let protocol_initializer = ProtocolInitializer::setup(
             self.protocol_parameters.clone().into(),
-            kes_secret_key_path,
+            kes_signer,
             signer_with_stake.kes_period,
             signer_with_stake.stake,
             rng,
@@ -123,26 +124,13 @@ impl SignerBuilder {
         ))
     }
 
-    /// Build non deterministic [SingleSigner] and [ProtocolInitializer] based on the registered parties.
-    pub fn build_single_signer(
-        &self,
-        signer_with_stake: SignerWithStake,
-        kes_secret_key_path: Option<&Path>,
-    ) -> StdResult<(SingleSigner, ProtocolInitializer)> {
-        self.build_single_signer_with_rng(
-            signer_with_stake,
-            kes_secret_key_path,
-            &mut rand_core::OsRng,
-        )
-    }
-
     /// Build deterministic [SingleSigner] and [ProtocolInitializer] based on the registered parties.
     ///
     /// Use for **TEST ONLY**.
     pub fn build_test_single_signer(
         &self,
         signer_with_stake: SignerWithStake,
-        kes_secret_key_path: Option<&Path>,
+        kes_signer: Option<Arc<dyn KesSigner>>,
     ) -> StdResult<(SingleSigner, ProtocolInitializer)> {
         let protocol_initializer_seed: [u8; 32] = signer_with_stake.party_id.as_bytes()[..32]
             .try_into()
@@ -150,7 +138,7 @@ impl SignerBuilder {
 
         self.build_single_signer_with_rng(
             signer_with_stake,
-            kes_secret_key_path,
+            kes_signer,
             &mut ChaCha20Rng::from_seed(protocol_initializer_seed),
         )
     }
@@ -186,7 +174,7 @@ mod test {
     use mithril_stm::RegisterError;
 
     use crate::{
-        crypto_helper::ProtocolRegistrationErrorWrapper,
+        crypto_helper::{KesSignerStandard, ProtocolRegistrationErrorWrapper},
         test_utils::{fake_data, MithrilFixtureBuilder},
     };
 
@@ -255,16 +243,23 @@ mod test {
             .build()
             .signers_fixture();
         let non_registered_signer = signers_from_another_fixture.first().unwrap();
+        let kes_signer = Some(Arc::new(KesSignerStandard::new(
+            non_registered_signer
+                .kes_secret_key_path()
+                .unwrap()
+                .to_path_buf(),
+            non_registered_signer
+                .operational_certificate_path()
+                .unwrap()
+                .to_path_buf(),
+        )) as Arc<dyn KesSigner>);
 
         let error = SignerBuilder::new(
             &fixture.signers_with_stake(),
             &fixture.protocol_parameters(),
         )
         .unwrap()
-        .build_test_single_signer(
-            non_registered_signer.signer_with_stake.clone(),
-            non_registered_signer.kes_secret_key_path(),
-        )
+        .build_test_single_signer(non_registered_signer.signer_with_stake.clone(), kes_signer)
         .expect_err(
             "We should not be able to construct a single signer from a not registered party",
         );
@@ -284,6 +279,10 @@ mod test {
         let fixture = MithrilFixtureBuilder::default().with_signers(3).build();
         let signers = fixture.signers_fixture();
         let signer = signers.first().unwrap();
+        let kes_signer = Some(Arc::new(KesSignerStandard::new(
+            signer.kes_secret_key_path().unwrap().to_path_buf(),
+            signer.operational_certificate_path().unwrap().to_path_buf(),
+        )) as Arc<dyn KesSigner>);
 
         let builder = SignerBuilder::new(
             &fixture.signers_with_stake(),
@@ -292,10 +291,7 @@ mod test {
         .unwrap();
 
         builder
-            .build_test_single_signer(
-                signer.signer_with_stake.clone(),
-                signer.kes_secret_key_path(),
-            )
+            .build_test_single_signer(signer.signer_with_stake.clone(), kes_signer)
             .expect("Should be able to build test single signer for a registered party");
     }
 
@@ -304,6 +300,10 @@ mod test {
         let fixture = MithrilFixtureBuilder::default().with_signers(3).build();
         let signers = fixture.signers_fixture();
         let signer = signers.first().unwrap();
+        let kes_signer = Some(Arc::new(KesSignerStandard::new(
+            signer.kes_secret_key_path().unwrap().to_path_buf(),
+            signer.operational_certificate_path().unwrap().to_path_buf(),
+        )) as Arc<dyn KesSigner>);
 
         let first_builder = SignerBuilder::new(
             &fixture.signers_with_stake(),
@@ -312,10 +312,7 @@ mod test {
         .unwrap();
 
         let (_, initializer) = first_builder
-            .build_test_single_signer(
-                signer.signer_with_stake.clone(),
-                signer.kes_secret_key_path(),
-            )
+            .build_test_single_signer(signer.signer_with_stake.clone(), kes_signer)
             .unwrap();
 
         let second_builder = SignerBuilder::new(